rename types to answers
