+++ turbo/generators/plugins/fix-deps-and-format.ts
@@ -3,7 +3,7 @@ import { $ } from '@repo/workspace-dependencies/zx'
 import { catchError, onProcSuccess } from '../helpers/proc'
 
 import type { PlopTypes } from '@turbo/gen'
-import type { Answers } from '../types'
+import type { Answers } from '../answers'
 
 export function fixDepsAndFormat(answers: Answers, _config: any, _plop: PlopTypes.NodePlopAPI) {
 	return new Promise((resolve, reject) => {
